code coverage

+++ tests/testthat/test-my_knn_cv.R
@@ -9,7 +9,7 @@ test_that("cross validation works", {
   penguins <- na.omit(penguins)
 
   x <- my_knn_cv(train = penguins, cl = 'species', k_nn = 3, k_cv = 5)[[1]]
-  expect_is(x, "numeric")
+  expect_is(x, "double")
 })
 
 test_that("non numeric parameter throws error", {
